requirements.txt
# Core
python-dotenv
numpy
pandas

# Computer Vision
opencv-python
ultralytics

# Hugging Face model download
huggingface_hub
fastapi
uvicorn
python-multipart
